--- scripts/concent_integration_tests/tests/test_mainnet_rpc.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-from scripts.concent_integration_tests.tests.playbooks import mainnet_rpc, run
-
-
-run.run_playbook(mainnet_rpc.MainnetRPCTest)  # type: ignore
