+++ scripts/concent_integration_tests/tests/test_rpc.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+from scripts.concent_integration_tests.tests.playbooks import rpc_test, run
+
+run.run_playbook(rpc_test.RPCTest)  # type: ignore
